Refactor RSS scraping functionality using Jsoup library
Removed the `RssUrl.java` file and replaced existing RSS scraping logic in `Main.java` with a more sophisticated process using the Jsoup library. This change was made to improve the efficiency and accuracy of parsing information from the RSS feed. The refactor also introduced a new Model `RssFeed.java` and a new service `RssScraper.java`. The `pom.xml` file was updated to include the Jsoup dependency.