Remove the Config class and simplify URL input
In this commit, the Config class from the utils package is removed. This was done because the application now uses a simpler method to get URLs - instead directly from a file named "input.txt". As a result, changes were also made in Main.java to reflect this new method of obtaining URLs, eliminating the previous use of the Config class. This change simplifies the codebase and makes URL input more straightforward.