Add Maven assembly plugin
Maven assembly plugin is added into the project to be able to create a single jar including its dependencies. The misc.xml file is updated to specify the resources of the project. Furthermore, pom.xml is updated to include the plugin and its configuration to specify the main class and file descriptor.