Add YAML config file and incorporate it into RssScraper
Added a snakeyaml dependency in the pom.xml and a YAML configuration file which is used to determine which elements should be scraped from an RSS feed.

The RssConfig class was also implemented to read from the YAML file and the RssScraper class has been adjusted to utilize this new class.

Finally, a new test case was added to verify the proper functionality of code changes with existing features. Updating the code to integrate this YAML configuration increases flexibility and extensibility of the scraping feature.