Add a configurable input file to the RSS scraper
The config.yml file has been updated to include an 'input-file' field so that users can specify the input file as per their requirements. A method was added to RssConfig.java to retrieve this filename from the configuration file. This allows for better flexibility and the user can change input files easily. Also optimized the test cases in ScraperTest to reduce redundant lines. The Main.java file has also been updated to use the new method in RssConfig.java, instead of the hardcoded 'input.txt'. This makes the application more flexible and user-friendly.