"Refactor RSS scraper to read URLs from config file"
This commit changes the `RssScraper` and `RssConfig` classes so that the URLs to be scraped are now read directly from the `config.yml` file instead of an `input.txt` file. This makes the source code cleaner and easier to maintain, as it removes the need to manage an extra file and reduces code redundancy. Furthermore, it makes it easier to add, remove or change URLs, as they are now all in a single centralized location. The `input.txt` file has been removed as it is no longer needed.